+++ db/migrate/20091028182228_create_price_field.rb
@@ -1,6 +1,6 @@
 class CreatePriceField < ActiveRecord::Migration
   def self.up
-    add_column :trainings, :price, :double
+    add_column :trainings, :price, :float
   end
 
   def self.down
